Add in NightHawk Theater
Add an abstract base class `Theater` that can generate `Screenings`
objects for the next two weeks. Build the `Nitehawk` theater off of this
base class. Add a function that takes all of the different `Theater`
subclasses, gets the screenings for all of them, and writes them to disc
as json, which can be called from poetry.